It seems the test failures were actually because the driver was turning on JMX monitoring by default! Also moved JmxPolicy to its own file.
