+++ build.gradle
@@ -39,7 +39,7 @@ buildscript {
     ext.typesafe_config_version = constants.getProperty("typesafeConfigVersion")
     ext.fileupload_version = '1.3.3'
     ext.junit_version = '4.12'
-    ext.mockito_version = '2.23.4'
+    ext.mockito_version = '2.18.3'
     ext.mockito_kotlin_version = '1.5.0'
     ext.hamkrest_version = '1.4.2.2'
     ext.jopt_simple_version = '5.0.2'
